+++ great_expectations/data_context/data_context/base_data_context.py
@@ -355,6 +355,7 @@ class BaseDataContext(EphemeralDataContext, ConfigPeer):
             self._data_context = CloudDataContext(
                 project_config=project_config,
                 runtime_environment=runtime_environment,
+                context_root_dir=context_root_dir,
                 ge_cloud_config=ge_cloud_config,
             )
         elif self._context_root_directory:
@@ -371,17 +372,16 @@ class BaseDataContext(EphemeralDataContext, ConfigPeer):
         # TODO: remove this method once refactor of DataContext is complete
         self._apply_temporary_overrides()
 
-        # Init plugin support
-        if self.plugins_directory is not None and os.path.exists(
-            self.plugins_directory
-        ):
-            sys.path.append(self.plugins_directory)
-
-        # We want to have directories set up before initializing usage statistics so that we can obtain a context instance id
-        self._in_memory_instance_id = (
-            None  # This variable *may* be used in case we cannot save an instance id
-        )
-
+        # # Init plugin support
+        # if self.plugins_directory is not None and os.path.exists(
+        #     self.plugins_directory
+        # ):
+        #     sys.path.append(self.plugins_directory)
+        #
+        # self._in_memory_instance_id = (
+        #     None  # This variable *may* be used in case we cannot save an instance id
+        # )
+        #
         # Init stores
         self._stores = {}
         self._init_stores(self.project_config_with_variables_substituted.stores)
@@ -627,15 +627,15 @@ class BaseDataContext(EphemeralDataContext, ConfigPeer):
         self.validation_operators[validation_operator_name] = new_validation_operator
         return new_validation_operator
 
-    def _normalize_absolute_or_relative_path(
-        self, path: Optional[str]
-    ) -> Optional[str]:
-        if path is None:
-            return
-        if os.path.isabs(path):
-            return path
-        else:
-            return os.path.join(self.root_directory, path)
+    # def _normalize_absolute_or_relative_path(
+    #     self, path: Optional[str]
+    # ) -> Optional[str]:
+    #     if path is None:
+    #         return
+    #     if os.path.isabs(path):
+    #         return path
+    #     else:
+    #         return os.path.join(self.root_directory, path)
 
     def _normalize_store_path(self, resource_store):
         if resource_store["type"] == "filesystem":
@@ -773,12 +773,12 @@ class BaseDataContext(EphemeralDataContext, ConfigPeer):
         ``great_expectations.yml`` is located."""
         return self._context_root_directory
 
-    @property
-    def plugins_directory(self):
-        """The directory in which custom plugin modules should be placed."""
-        return self._normalize_absolute_or_relative_path(
-            self.project_config_with_variables_substituted.plugins_directory
-        )
+    # @property
+    # def plugins_directory(self):
+    #     """The directory in which custom plugin modules should be placed."""
+    #     return self._normalize_absolute_or_relative_path(
+    #         self.project_config_with_variables_substituted.plugins_directory
+    #     )
 
     @property
     def usage_statistics_handler(self) -> Optional[UsageStatisticsHandler]:
@@ -929,9 +929,9 @@ class BaseDataContext(EphemeralDataContext, ConfigPeer):
             self._in_memory_instance_id = instance_id
         return instance_id
 
-    @property
-    def config(self) -> DataContextConfig:
-        return self._project_config
+    # @property
+    # def config(self) -> DataContextConfig:
+    #     return self._project_config
 
     #####
     #
@@ -939,44 +939,44 @@ class BaseDataContext(EphemeralDataContext, ConfigPeer):
     #
     #####
 
-    def get_config_with_variables_substituted(
-        self, config: Optional[DataContextConfig] = None
-    ) -> DataContextConfig:
-        """
-        Substitute vars in config of form ${var} or $(var) with values found in the following places,
-        in order of precedence: ge_cloud_config (for Data Contexts in GE Cloud mode), runtime_environment,
-        environment variables, config_variables, or ge_cloud_config_variable_defaults (allows certain variables to
-        be optional in GE Cloud mode).
-        """
-        if not config:
-            config = self.config
-
-        substitutions: dict = self._determine_substitutions()
-
-        if self.ge_cloud_mode:
-            ge_cloud_config_variable_defaults = {
-                "plugins_directory": self._normalize_absolute_or_relative_path(
-                    DataContextConfigDefaults.DEFAULT_PLUGINS_DIRECTORY.value
-                ),
-                "usage_statistics_url": DEFAULT_USAGE_STATISTICS_URL,
-            }
-            for config_variable, value in ge_cloud_config_variable_defaults.items():
-                if substitutions.get(config_variable) is None:
-                    logger.info(
-                        f'Config variable "{config_variable}" was not found in environment or global config ('
-                        f'{self.GLOBAL_CONFIG_PATHS}). Using default value "{value}" instead. If you would '
-                        f"like to "
-                        f"use a different value, please specify it in an environment variable or in a "
-                        f"great_expectations.conf file located at one of the above paths, in a section named "
-                        f'"ge_cloud_config".'
-                    )
-                    substitutions[config_variable] = value
-
-        return DataContextConfig(
-            **substitute_all_config_variables(
-                config, substitutions, self.DOLLAR_SIGN_ESCAPE_STRING
-            )
-        )
+    # def get_config_with_variables_substituted(
+    #     self, config: Optional[DataContextConfig] = None
+    # ) -> DataContextConfig:
+    #     """
+    #     Substitute vars in config of form ${var} or $(var) with values found in the following places,
+    #     in order of precedence: ge_cloud_config (for Data Contexts in GE Cloud mode), runtime_environment,
+    #     environment variables, config_variables, or ge_cloud_config_variable_defaults (allows certain variables to
+    #     be optional in GE Cloud mode).
+    #     """
+    #     if not config:
+    #         config = self.config
+    #
+    #     substitutions: dict = self._determine_substitutions()
+    #
+    #     if self.ge_cloud_mode:
+    #         ge_cloud_config_variable_defaults = {
+    #             "plugins_directory": self._normalize_absolute_or_relative_path(
+    #                 DataContextConfigDefaults.DEFAULT_PLUGINS_DIRECTORY.value
+    #             ),
+    #             "usage_statistics_url": DEFAULT_USAGE_STATISTICS_URL,
+    #         }
+    #         for config_variable, value in ge_cloud_config_variable_defaults.items():
+    #             if substitutions.get(config_variable) is None:
+    #                 logger.info(
+    #                     f'Config variable "{config_variable}" was not found in environment or global config ('
+    #                     f'{self.GLOBAL_CONFIG_PATHS}). Using default value "{value}" instead. If you would '
+    #                     f"like to "
+    #                     f"use a different value, please specify it in an environment variable or in a "
+    #                     f"great_expectations.conf file located at one of the above paths, in a section named "
+    #                     f'"ge_cloud_config".'
+    #                 )
+    #                 substitutions[config_variable] = value
+    #
+    #     return DataContextConfig(
+    #         **substitute_all_config_variables(
+    #             config, substitutions, self.DOLLAR_SIGN_ESCAPE_STRING
+    #         )
+    #     )
 
     def escape_all_config_variables(
         self,
