+++ great_expectations/data_context/data_context/cloud_data_context.py
@@ -1,13 +1,20 @@
 import logging
-from typing import Mapping, Union
+import os
+from typing import Mapping, Optional, Union
 
 from great_expectations.data_context.data_context.abstract_data_context import (
     AbstractDataContext,
 )
-from great_expectations.data_context.types.base import DataContextConfig, GeCloudConfig
+from great_expectations.data_context.types.base import (
+    DEFAULT_USAGE_STATISTICS_URL,
+    DataContextConfig,
+    DataContextConfigDefaults,
+    GeCloudConfig,
+)
 from great_expectations.data_context.types.data_context_variables import (
     CloudDataContextVariables,
 )
+from great_expectations.data_context.util import substitute_all_config_variables
 
 logger = logging.getLogger(__name__)
 
@@ -22,6 +29,7 @@ class CloudDataContext(AbstractDataContext):
         project_config: Union[DataContextConfig, Mapping],
         runtime_environment: dict,
         ge_cloud_config: GeCloudConfig,
+        context_root_dir: Optional[str] = None,
     ) -> None:
         """
         CloudDataContext constructor
@@ -34,6 +42,7 @@ class CloudDataContext(AbstractDataContext):
         """
         super().__init__(runtime_environment=runtime_environment)
         self._ge_cloud_mode = True  # property needed for backward compatibility
+        self._context_root_directory = context_root_dir
         self._ge_cloud_config = ge_cloud_config
         self._project_config = self._apply_global_config_overrides(
             config=project_config
@@ -41,3 +50,65 @@ class CloudDataContext(AbstractDataContext):
 
     def _init_variables(self) -> CloudDataContextVariables:
         raise NotImplementedError
+
+    def get_config_with_variables_substituted(
+        self, config: Optional[DataContextConfig] = None
+    ) -> DataContextConfig:
+        """
+        Substitute vars in config of form ${var} or $(var) with values found in the following places,
+        in order of precedence: ge_cloud_config (for Data Contexts in GE Cloud mode), runtime_environment,
+        environment variables, config_variables, or ge_cloud_config_variable_defaults (allows certain variables to
+        be optional in GE Cloud mode).
+        """
+        if not config:
+            config = self.config
+
+        substitutions: dict = self._determine_substitutions()
+
+        ge_cloud_config_variable_defaults = {
+            "plugins_directory": self._normalize_absolute_or_relative_path(
+                DataContextConfigDefaults.DEFAULT_PLUGINS_DIRECTORY.value
+            ),
+            "usage_statistics_url": DEFAULT_USAGE_STATISTICS_URL,
+        }
+        for config_variable, value in ge_cloud_config_variable_defaults.items():
+            if substitutions.get(config_variable) is None:
+                logger.info(
+                    f'Config variable "{config_variable}" was not found in environment or global config ('
+                    f'{self.GLOBAL_CONFIG_PATHS}). Using default value "{value}" instead. If you would '
+                    f"like to "
+                    f"use a different value, please specify it in an environment variable or in a "
+                    f"great_expectations.conf file located at one of the above paths, in a section named "
+                    f'"ge_cloud_config".'
+                )
+                substitutions[config_variable] = value
+
+        return DataContextConfig(
+            **substitute_all_config_variables(
+                config, substitutions, self.DOLLAR_SIGN_ESCAPE_STRING
+            )
+        )
+
+    @property
+    def root_directory(self):
+        """The root directory for configuration objects in the data context; the location in which
+        ``great_expectations.yml`` is located."""
+        return self._context_root_directory
+
+    def _normalize_absolute_or_relative_path(
+        self, path: Optional[str]
+    ) -> Optional[str]:
+        """
+        Shared by CloudDataContext and FileDAtaContext
+        Args:
+            path ():
+
+        Returns:
+
+        """
+        if path is None:
+            return
+        if os.path.isabs(path):
+            return path
+        else:
+            return os.path.join(self.root_directory, path)
